// FrontEnd/BillionShop/src/app/rcv/model/rcv00101_Head.ts
export interface Rcv00101_Head {
    riId: Number;
    riDate: String;
    riCode: String;
    riGenUser: String;
    riStatus: String;
}

export interface Rcv001_Head {
    riId: Number;
    riDate: String;
    riCode: String;
    riGenUser: String;
    riStatus: String;
}

export interface RcvId{
    rcvId:number;
}